d2: Fix apps2sd to usb storage
While this may not be a very good idea to use (do you really want
to have a usb stick sticking out of your phone??), the current
fstab is incorrect and needs to be fixed.

Removable storage is always noemulatedsd (and defined as such
in storage_list.xml).  The fstab should be consistent with the
rest of android.

Conflicts:
	rootdir/etc/fstab.qcom

Change-Id: I5510faa54d29df9861bc9f699f50c5a3074e9e6d